+++ src/Collection.php
@@ -696,22 +696,6 @@ class Collection implements \Countable
         return $documents;
     }
 
-    /**
-     * Save document
-     *
-     * @deprecated since v.1.8.0 use Document::save() method
-     * @param \Sokil\Mongo\Document $document
-     * @param bool $validate validate or not before save
-     * @return \Sokil\Mongo\Collection
-     * @throws \Sokil\Mongo\Exception
-     * @throws \Sokil\Mongo\Document\InvalidDocumentException
-     */
-    public function saveDocument(Document $document, $validate = true)
-    {
-        $document->save($validate);
-        return $this;
-    }
-
     public function deleteDocument(Document $document)
     {
         if($document->triggerEvent('beforeDelete')->isCancelled()) {